requirements.txt
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
bcrypt==4.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.16.0
idna==3.11
passlib==1.7.4
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.6
uvicorn==0.24.0
